% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{any_duplicated_linter}
\alias{any_duplicated_linter}
\title{Require usage of \code{anyDuplicated(x) > 0} over \code{any(duplicated(x))}}
\usage{
any_duplicated_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/any_duplicated_linter}.
}
