% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{any_is_na_linter}
\alias{any_is_na_linter}
\title{Require usage of \code{anyNA(x)} over \code{any(is.na(x))}}
\usage{
any_is_na_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/any_is_na_linter}.
}
