% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{class_equals_linter}
\alias{class_equals_linter}
\title{Block comparison of class with \code{==}}
\usage{
class_equals_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/class_equals_linter}.
}
