% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{condition_message_linter}
\alias{condition_message_linter}
\title{Block usage of \code{paste()} and \code{paste0()} with messaging functions using \code{...}}
\usage{
condition_message_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/condition_message_linter}.
}
