% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{duplicate_argument_linter}
\alias{duplicate_argument_linter}
\title{Duplicate argument linter}
\usage{
duplicate_argument_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/duplicate_argument_linter}.
}
