% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_comparison_linter}
\alias{expect_comparison_linter}
\title{Require usage of \code{expect_gt(x, y)} over \code{expect_true(x > y)} (and similar)}
\usage{
expect_comparison_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_comparison_linter}.
}
