% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_identical_linter}
\alias{expect_identical_linter}
\title{Require usage of \code{expect_identical(x, y)} where appropriate}
\usage{
expect_identical_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_identical_linter}.
}
