% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_named_linter}
\alias{expect_named_linter}
\title{Require usage of \code{expect_named(x, n)} over \code{expect_equal(names(x), n)}}
\usage{
expect_named_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_named_linter}.
}
