% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_not_linter}
\alias{expect_not_linter}
\title{Require usage of \code{expect_false(x)} over \code{expect_true(!x)}}
\usage{
expect_not_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_not_linter}.
}
