% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_null_linter}
\alias{expect_null_linter}
\title{Require usage of \code{expect_null} for checking \code{NULL}}
\usage{
expect_null_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_null_linter}.
}
