% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_true_false_linter}
\alias{expect_true_false_linter}
\title{Require usage of \code{expect_true(x)} over \code{expect_equal(x, TRUE)}}
\usage{
expect_true_false_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_true_false_linter}.
}
