% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{expect_type_linter}
\alias{expect_type_linter}
\title{Require usage of \code{expect_type(x, type)} over \code{expect_equal(typeof(x), type)}}
\usage{
expect_type_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/expect_type_linter}.
}
