\name{flint-defunct}
\alias{flint-defunct}
\title{Defunct Functions in Package \pkg{flint}}
\description{
  The functions and other objects listed here are no longer part of
  \pkg{flint} as they are no longer needed.
}
\usage{
## Nothing yet!
}
\details{
  These either are stubs reporting that they are defunct or have been
  removed completely (apart from being documented here).
}
\seealso{
  \code{\link{Deprecated}}, \code{\link{base-deprecated}},
  \code{\link{flint-deprecated}}, \code{\link{flint-notyet}}.
}
\keyword{internal}
