\name{flint-deprecated}
\alias{flint-deprecated}
\title{Deprecated Functions in Package \pkg{flint}}
\description{
  The functions and other objects listed here are provided only for
  compatibility with older versions of \pkg{flint} and may become
  defunct as soon as the next release.
}
\usage{
## Nothing yet!
}
\seealso{
  \code{\link{Defunct}}, \code{\link{base-defunct}},
  \code{\link{flint-defunct}}, \code{\link{flint-notyet}}.
}
\keyword{internal}
