% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{function_return_linter}
\alias{function_return_linter}
\title{Lint common mistakes/style issues cropping up from return statements}
\usage{
function_return_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/function_return_linter}.
}
