% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{is_numeric_linter}
\alias{is_numeric_linter}
\title{Redirect \code{is.numeric(x) || is.integer(x)} to just use \code{is.numeric(x)}}
\usage{
is_numeric_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/is_numeric_linter}.
}
