% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{length_levels_linter}
\alias{length_levels_linter}
\title{Require usage of nlevels over length(levels(.))}
\usage{
length_levels_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/length_levels_linter}.
}
