% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{list_comparison_linter}
\alias{list_comparison_linter}
\title{Block usage of comparison operators with known-list() functions like lapply}
\usage{
list_comparison_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/list_comparison_linter}.
}
