% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-linters.R
\name{list_linters}
\alias{list_linters}
\title{Get the list of linters in \code{flint}}
\usage{
list_linters()
}
\value{
A character vector
}
\description{
Get the list of linters in \code{flint}
}
\examples{
list_linters()
}
