% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{literal_coercion_linter}
\alias{literal_coercion_linter}
\title{Require usage of correctly-typed literals over literal coercions}
\usage{
literal_coercion_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/literal_coercion_linter}.
}
