% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{matrix_apply_linter}
\alias{matrix_apply_linter}
\title{Require usage of \code{colSums(x)} or \code{rowSums(x)} over \code{apply(x, ., sum)}}
\usage{
matrix_apply_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/matrix_apply_linter}.
}
