% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{missing_argument_linter}
\alias{missing_argument_linter}
\title{Missing argument linter}
\usage{
missing_argument_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/missing_argument_linter}.
}
