% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{numeric_leading_zero_linter}
\alias{numeric_leading_zero_linter}
\title{Require usage of a leading zero in all fractional numerics}
\usage{
numeric_leading_zero_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/numeric_leading_zero_linter}.
}
