% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{outer_negation_linter}
\alias{outer_negation_linter}
\title{Require usage of \code{!any(x)} over \code{all(!x)}, \code{!all(x)} over \code{any(!x)}}
\usage{
outer_negation_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/outer_negation_linter}.
}
