% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{paste_linter}
\alias{paste_linter}
\title{Raise lints for several common poor usages of \code{paste()}}
\usage{
paste_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/paste_linter}.
}
