% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{pipe_consistency_linter}
\alias{pipe_consistency_linter}
\title{Pipe consistency linter}
\usage{
pipe_consistency_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/pipe_consistency_linter}.
}
