% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{redundant_equals_linter}
\alias{redundant_equals_linter}
\title{Block usage of \code{==}, \code{!=} on logical vectors}
\usage{
redundant_equals_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/redundant_equals_linter}.
}
