% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{redundant_ifelse_linter}
\alias{redundant_ifelse_linter}
\title{Prevent \code{ifelse()} from being used to produce \code{TRUE}/\code{FALSE} or \code{1}/\code{0}}
\usage{
redundant_ifelse_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/redundant_ifelse_linter}.
}
