% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{rep_len_linter}
\alias{rep_len_linter}
\title{Require usage of rep_len(x, n) over rep(x, length.out = n)}
\usage{
rep_len_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/rep_len_linter}.
}
