% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{sample_int_linter}
\alias{sample_int_linter}
\title{Require usage of sample.int(n, m, ...) over sample(1:n, m, ...)}
\usage{
sample_int_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/sample_int_linter}.
}
