% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_flint}
\alias{setup_flint}
\title{Setup flint}
\usage{
setup_flint()
}
\value{
Imports files necessary for \code{flint} to work but doesn't return any
value in R.
}
\description{
This stores the default rules and internal files in \code{inst/flint}. It also
imports \code{sgconfig.yml} that is used by \code{ast-grep}. This file must live at the
root of the project and cannot be renamed.
}
