% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_tinylint}
\alias{setup_tinylint}
\title{Setup tinylint}
\usage{
setup_tinylint()
}
\value{
Imports files necessary for \code{tinylint} to work but doesn't return any
value in R.
}
\description{
This stores the default rules and internal files in \code{inst/tinylint}. It also
imports \code{sgconfig.yml} that is used by \code{ast-grep}. This file must live at the
root of the project and cannot be renamed.
}
