% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{stopifnot_all_linter}
\alias{stopifnot_all_linter}
\title{Block usage of all() within stopifnot()}
\usage{
stopifnot_all_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/stopifnot_all_linter}.
}
