% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{T_and_F_symbol_linter}
\alias{T_and_F_symbol_linter}
\title{\code{T} and \code{F} symbol linter}
\usage{
T_and_F_symbol_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/T_and_F_symbol_linter}.
}
