% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylint-package.R
\docType{package}
\name{tinylint-package}
\alias{tinylint}
\alias{tinylint-package}
\title{tinylint: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etiennebacher/tinylint}
  \item Report bugs at \url{https://github.com/etiennebacher/tinylint/issues}
}

}
\keyword{internal}
