% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{undesirable_function_linter}
\alias{undesirable_function_linter}
\title{Undesirable function linter}
\usage{
undesirable_function_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/undesirable_function_linter}.
}
