% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{undesirable_operator_linter}
\alias{undesirable_operator_linter}
\title{Undesirable operator linter}
\usage{
undesirable_operator_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/undesirable_operator_linter}.
}
