% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{unnecessary_nesting_linter}
\alias{unnecessary_nesting_linter}
\title{Block instances of unnecessary nesting}
\usage{
unnecessary_nesting_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/unnecessary_nesting_linter}.
}
