% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters_factory.R
\name{unreachable_code_linter}
\alias{unreachable_code_linter}
\title{Block unreachable code and comments following return statements}
\usage{
unreachable_code_linter
}
\description{
See \url{https://lintr.r-lib.org/reference/unreachable_code_linter}.
}
