% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_flint}
\alias{update_flint}
\title{Update the \code{flint} setup}
\usage{
update_flint(path = ".")
}
\arguments{
\item{path}{Path to package or project root.}
}
\value{
Can add new files in the \code{flint/rules} folder, doesn't return anything.
}
\description{
When \code{flint} is updated, it can ship new built-in rules or update existing
ones. \code{update_flint()} will automatically add those new rules to the
\code{flint/rules/builtin} folder. Custom rules stored in \code{flint/rules/custom}
are not affected.
}
\examples{
\dontrun{
  update_flint()
}
}
