% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylint-package.R
\docType{package}
\name{flint-package}
\alias{flint}
\alias{flint-package}
\title{flint: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etiennebacher/flint}
  \item Report bugs at \url{https://github.com/etiennebacher/flint/issues}
}

}
\keyword{internal}
