% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylint-package.R
\docType{package}
\name{flint-package}
\alias{flint}
\alias{flint-package}
\title{flint: Find and Fix Lints in R Code}
\description{
Lints are code patterns that are not optimal because they are inefficient, forget corner cases, or less readable. 'flint' provides a small set of functions to detect those lints and automatically fix them. It builds on 'astgrepr', which itself uses the Rust crate 'ast-grep' to parse and navigate R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://flint.etiennebacher.com}
  \item \url{https://github.com/etiennebacher/flint}
  \item Report bugs at \url{https://github.com/etiennebacher/flint/issues}
}

}
\author{
\strong{Maintainer}: Etienne Bacher \email{etienne.bacher@protonmail.com} [copyright holder]

}
\keyword{internal}
