\name{flint-package}
\alias{flint}
\alias{flint-package}
\title{\R{} Package \pkg{flint}}
\description{
An \R{} interface to the \abbr{FLINT} (Fast Library for Number Theory)
C library, an extension of \abbr{GNU} \abbr{MFPR} (Multiple Precision
Floating-Point Reliable Library) and \abbr{GNU} \abbr{MP} (Multiple
Precision Arithmetic Library).
}
\details{
To render a list of available help topics, use
\code{\link{help}(package = "flint")}.

To report a bug or request a change, use
\code{\link{bug.report}(package = "flint")}.
}
\author{
Mikael Jagan
\email{jaganmn@mcmaster.ca}
}
\docType{package}
