% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint.R
\name{lint}
\alias{lint}
\alias{lint_diff}
\alias{lint_text}
\title{List all lints in a file or a directory}
\usage{
lint(path = ".", linters = NULL, open = TRUE)

lint_diff(path = ".", open = TRUE)

lint_text(text)
}
\arguments{
\item{path}{A valid path to a file or a directory. Relative paths are accepted.}

\item{linters}{A character vector with the names of the rules to apply. See
the entire list of rules with \code{list_linters()}.}

\item{open}{If \code{TRUE} (default) and if this is used in the RStudio IDE, lints
will be shown with markers.}
}
\value{
A dataframe where each row is a lint. The columns show the text, its
location (both the position in the text and the file in which it was found)
and the severity.
}
\description{
\code{lint()}, \code{lint_text()} and \code{lint_diff()} all produce a list of lints. The
only difference is in the input they take:
\itemize{
\item \code{lint()} takes path to files or directories
\item \code{lint_text()} takes some text input
\item \code{lint_diff()} takes a path to a directory but only looks at files that have
changed since the last commit.
}
}
