% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{update_flint}
\alias{update_flint}
\title{Update the \code{flint} setup}
\usage{
update_flint(path = ".")
}
\arguments{
\item{path}{Path to package or project root.}
}
\value{
Can add new files in the \code{flint/rules} folder, doesn't return anything.
}
\description{
When \code{flint} is updated, it can ship new built-in rules. \code{update_flint()} will
automatically add those new rules in the \code{flint/rules} folder. New rules are
only determined by their names and rules that already exist in \code{flint/rules}
are not affected.

For instance, if you added a custom rule \code{use_paste.yml}, then it will never
be removed by \code{update_flint()}, even if \code{flint} later adds a built-in rule
also named \code{use_paste.yml}.
}
\examples{
\dontrun{
  update_flint()
}
}
